Feature/monitor park

+++ webapp/backend/src/main/java/snappark/backend/entity/Park.java
@@ -44,7 +44,7 @@ public class Park {
     private float longitude;
 
     @Column(nullable = false)
-    private float entranceFee;
+    private double entranceFee;
 
     @Column
     private int maxLotation;
